requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
slowapi==0.1.9
chromadb>=1.4.1
sentence-transformers==2.7.0
transformers==4.36.2
torch==2.1.2
pandas==2.1.4
numpy~=1.26.4
Pillow==10.2.0
pytesseract==0.3.10
python-multipart==0.0.6
deep-translator==1.11.4
gTTS==2.5.0
google-generativeai==0.3.2
google-cloud-translate==3.14.0
twilio==8.11.1
python-dotenv==1.0.0
langdetect==1.0.9
